[Refactoring] WriteModule Spy 적용
정상동작 테스트를 위해 Spy로 변경
